requirements.txt
streamlit
PyMuPDF
PyPDF2
Pillow
